refactor: Layers, Components 디렉토리 구조화 [#98]
CoworkerStatus를 Layers 폴더로 분리
