+++ client/src/components/CoworkerStatusItem/Avator/index.tsx
